requirements.txt
pyyaml
newsapi-python